* change build to 32/64bit ppc+i386. We'll continue building for PPC, but supporting it is no longer a priority. Intel 32/64 will be primary focus the * fix a few references to plain int/unsigned in CRDTabView which were causing warnings on 64 bit. NSUInteger and NSInteger should be used wherever feasible * fixes to server filtering: 	- now works with both label and hostname; 	- now works correctly when no results are found; 	- dragging no longer results in weirdness; 	- performance improvement (call listUpdated only once in -[AppController filterServers]) * Replace most of CRDSession's accessors with objc2.0 synthesized accessors * Replace a couple of ints with NSUInteger/NSInteger, where appropriate and side-affect free (some ints need to be plain ints, due to reliance on things like scanInt:) * Replace a couple NSEnumerator references with fast enumeration (for (x in a) {} ). Fast enumeration is definitely a good thing for performance and readability, and should be used (for new and existing NSEnumerator code)

+++ Source/CRDTabView.h
@@ -37,15 +37,15 @@
 - (IBAction)selectNextItem:(id)sender;
 - (IBAction)selectPreviousItem:(id)sender;
 - (IBAction)selectItem:(id)item;
-- (IBAction)selectItemAtIndex:(unsigned)index;
+- (IBAction)selectItemAtIndex:(NSUInteger)index;
 
-- (id)itemAtIndex:(unsigned)index;
+- (id)itemAtIndex:(NSUInteger)index;
 - (id)selectedItem;
 
-- (unsigned)indexOfItem:(id)item;
-- (unsigned)indexOfSelectedItem;
+- (NSUInteger)indexOfItem:(id)item;
+- (NSUInteger)indexOfSelectedItem;
 
-- (unsigned)numberOfItems;
+- (NSUInteger)numberOfItems;
 
 - (void)addItem:(id)item;
 - (void)removeItem:(id)item;
